[PYOKEMON-152] [Feature/#51] TicketPage를 MyBookingsPage로 이름 변경
